+++ lib/custom_widgets.dart
@@ -14,6 +14,10 @@ import 'package:fl_chart/fl_chart.dart';
 
 import 'package:flutter_multi_formatter/flutter_multi_formatter.dart';
 
+import 'package:logger/logger.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'package:flutter/foundation.dart' as foundation;
+
 import './custom_widgets/app_style.dart';
 import './custom_widgets/helper.dart';
 import './dashboard/daily_info_model.dart';
@@ -45,6 +49,8 @@ part './custom_widgets/clear_icon.dart';
 part './custom_widgets/custom_expandable_widget.dart';
 
 part './dashboard/mini_information_widget.dart';
+
+part './helper/Helper.dart';